fix:modify,delete button

+++ view.php
@@ -35,7 +35,7 @@ mysqli_query($connect, $update_query);
     <div class="d-flex justify-content-center mt-4">
         <a href="./main.php" class="btn btn-outline-dark me-3">목록</a>
         <?php
-        if(isset($_SESSION['userid'])){
+        if(isset ($_SESSION['userid']) && $_SESSION['userid'] === $rows['id']){
             echo '<a href="./modify.php?number=' . $number . '&id=' . $_rows['id'] . '" class="btn btn-outline-dark me-3">수정</a>';
             echo '<a href="./board_delete.php?number=' . $number. '&id='. $_rows['id'] . '" class="btn btn-outline-danger">삭제</a>';
         }
